+++ src/Sinch/Conversation/EventDestinations/CreateWebhookRequest.cs
@@ -1,12 +1,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Sinch.Conversation.Webhooks
+namespace Sinch.Conversation.EventDestinations
 {
-    /// <summary>
-    ///     Represents a destination for receiving callbacks from the Conversation API.
-    /// </summary>
-    public sealed class Webhook
+    public sealed class CreateWebhookRequest
     {
         /// <summary>
         ///     Gets or sets the target type.
@@ -16,17 +13,12 @@ namespace Sinch.Conversation.Webhooks
         /// <summary>
         ///     The app that this webhook belongs to.
         /// </summary>
-        public string? AppId { get; init; }
+        public required string AppId { get; init; }
 
         /// <summary>
         ///     Gets or sets the client credentials.
         /// </summary>
         public ClientCredentials? ClientCredentials { get; init; }
-
-        /// <summary>
-        ///     Gets or sets the ID of the webhook.
-        /// </summary>
-        public string? Id { get; init; }
 
         /// <summary>
         ///     Optional secret to be used to sign contents of webhooks sent by the Conversation API.
@@ -57,7 +49,6 @@ namespace Sinch.Conversation.Webhooks
             sb.Append("class Webhook {\n");
             sb.Append("  AppId: ").Append(AppId).Append("\n");
             sb.Append("  ClientCredentials: ").Append(ClientCredentials).Append("\n");
-            sb.Append("  Id: ").Append(Id).Append("\n");
             sb.Append("  Target: ").Append(Target).Append("\n");
             sb.Append("  TargetType: ").Append(TargetType).Append("\n");
             sb.Append("  Triggers: ").Append(Triggers).Append("\n");
